+++ R/dmdprices-package.R
@@ -50,5 +50,6 @@
 #' @importFrom cli cli_abort cli_inform cli_progress_step cli_warn
 #' @importFrom stringdist stringdist
 #' @importFrom tibble tibble tribble
+#' @importFrom memoise memoise
 ## usethis namespace: end
 NULL